+++ recipes/yum_remove.rb
@@ -1,11 +1,10 @@
-Chef::Log.info node['sls_utils']['yum_remove']
-
 node['sls_utils']['yum_remove'].each do |pkg|
+    Chef::Log.info "Removing package: #{pkg}"
 
-    Chef::Log.info pkg
-
-    package "remove-#{pkg}" do
+    p = package "remove-#{pkg}" do
         package_name pkg
-        action :remove
+        action :nothing
     end
+
+    p.run_action(:remove)
 end